+++ client/smart-waterloo-web/src/data/Events.ts
