+++ client/smart-waterloo-web/src/pages/About/About.css
@@ -1,3 +1,7 @@
 .AboutHeader{
 	margin-bottom: 1rem;
 }
+.aboutUsImage{
+	width: 100%;
+	height: auto;
+}
