+++ client/smart-waterloo-web/src/pages/CreateOrg/index.tsx
@@ -1,0 +1,1 @@
+export {default} from "./CreateOrg"
