+++ client/smart-waterloo-web/src/pages/SignUp/MeetAvatar/MeetAvatar.tsx
@@ -1,5 +1,6 @@
 // import {Link} from "react-router-dom";
 // import Avatar from "../../../images/fullAvatar.png";
+import { randomString } from "../../../modules/randomData";
 import "./MeetAvatar.css";
 
 //Todo change buttons to links
@@ -10,16 +11,6 @@ type MeetAvatarProps = {
 	updateParentState: Function
 	avatarData: {avatarString:string}
 };
-const randomString = () => {
-    var result           = '';
-    var characters       = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    var charactersLength = characters.length;
-    for ( var i = 0; i < 8; i++ ) {
-      	result += characters.charAt(Math.floor(Math.random() * charactersLength));
-   	}
-	console.log(result);
-   	return result;
-}
 function MeetAvatar(props:MeetAvatarProps) {
 	const spacing = {
 		margin: "0.5rem 0"
