+++ client/smart-waterloo-web/src/pages/Surveys/Surveys.tsx
@@ -3,11 +3,11 @@ import Navbar from "../../components/Navbar";
 import { MobileContext, IdContext, AddressContext } from "../../App";
 import {useContext, useState} from "react";
 import "./Surveys.css";
-import {exampleSurveys} from "../data/Surveys";
+import {exampleSurveys} from "../../data/Surveys";
 import SurveyPanel from "./SurveyPanel";
 import { useNavigate } from "react-router-dom";
-import { defaultUserData} from "../data/Users";
-import { getUserData} from "../data/getData"
+import { defaultUserData} from "../../data/Users";
+import { getUserData} from "../../data/getData"
 
 
 const Surveys = (props: {org: boolean}) => {
